+++ src/main/java/com/commercetools/sync/products/ProductSyncOptionsBuilder.java
@@ -71,7 +71,6 @@ public final class ProductSyncOptionsBuilder
             errorCallback,
             warningCallback,
             batchSize,
-            allowUuid,
             syncFilter,
             beforeUpdateCallback,
             beforeCreateCallback,