+++ src/main/java/com/commercetools/sync/producttypes/ProductTypeSyncOptionsBuilder.java
@@ -40,7 +40,6 @@ public final class ProductTypeSyncOptionsBuilder extends BaseSyncOptionsBuilder<
             errorCallback,
             warningCallback,
             batchSize,
-            allowUuid,
             beforeUpdateCallback,
             beforeCreateCallback
         );